MVP-5588+5590-4: Restructure (file-structure) notifi-node-sample
