--- packages/notifi-node-sample/tsconfig.build.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "compilerOptions": {
-    "outDir": "./dist"
-  },
-  "include": [
-    "lib"
-  ],
-  "types": ["express"],
-}
