+++ packages/notifi-node-sample/types.d.ts
@@ -1,0 +1,7 @@
+import { Request } from 'express';
+
+declare module 'express' {
+  interface Request {
+    _startTime?: number;
+  }
+}
